+++ changelog.md
@@ -18,3 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Function gt can now extract from a mongo engine Document
 - Elasticsearch and mongoengine are not requirements now
+
+## [0.0.21] - 13-01-2020
+### Added
+- [Await Process Decorator](https://github.com/iurii-ebs/drf-util/commit/fbe693d3b5aabedc0594a8c2c6f5f346a8091b4d#diff-6ccbb3dab6897856ba86fe421ccd584bR73)
+- [Browser Headers](https://github.com/iurii-ebs/drf-util/commit/fbe693d3b5aabedc0594a8c2c6f5f346a8091b4d#diff-ab3a5ae228e9c73ee208114a62856078R3)